Timeline with database connection
Refactoring and cleanup
