+++ src/app/api/quality_matrix/timeline.py
@@ -1,22 +1,10 @@
-import os
 from typing import Mapping
-from urllib.parse import quote
 
 from databases import Database
-from sqlalchemy import MetaData, create_engine, inspect, select
+from sqlalchemy import select
 
-from app.api.quality_matrix.models import Timeline, timeline_table
-
-
-def database_url():
-    POSTGRES_USER: str = os.getenv("POSTGRES_USER")
-    POSTGRES_PASSWORD = quote(os.getenv("POSTGRES_PASSWORD"))
-    POSTGRES_SERVER: str = os.getenv("POSTGRES_SERVER", "localhost")
-    POSTGRES_PORT: str = os.getenv(
-        "POSTGRES_PORT", 5432
-    )  # default postgres port is 5432
-    POSTGRES_DB: str = os.getenv("POSTGRES_DB", "tdd")
-    return f"postgresql://{POSTGRES_USER}:{POSTGRES_PASSWORD}@{POSTGRES_SERVER}:{POSTGRES_PORT}/{POSTGRES_DB}"
+from app.api.quality_matrix.models import Timeline
+from app.db.core import create_timeline_table, has_table
 
 
 async def timestamps(database: Database):
@@ -30,22 +18,3 @@ async def timestamps(database: Database):
     if result is None:
         return []
     return [entry.timestamp for entry in result]
-
-
-async def has_table(name: str):
-    engine = create_engine(database_url())
-    ins = inspect(engine)
-    return ins.dialect.has_table(engine.connect(), name)
-
-
-async def create_timeline_table():
-    engine, table = await get_table()
-    with engine.connect():
-        table.create()
-
-
-async def get_table():
-    engine = create_engine(database_url())
-    meta = MetaData(engine)
-    table = timeline_table(meta)
-    return engine, table
