+++ src/app/core/tasks.py
@@ -2,7 +2,7 @@ from typing import Callable
 
 from fastapi import FastAPI
 
-from app.pg.tasks import close_db_connection, connect_to_db
+from app.db.tasks import close_db_connection, connect_to_db
 
 
 def create_start_app_handler(app: FastAPI) -> Callable:
